제품 입고 관리, 현황까지 완료

+++ src/main/java/com/example/pjymes/service/productManage/ProductWarehousingService.java
@@ -8,7 +8,7 @@ import java.util.List;
 
 public interface ProductWarehousingService {
 
-    Long register(List<ProductWarehousingDTO> productWarehousingDTOS);
+    int register(List<ProductWarehousingDTO> productWarehousingDTOS);
 
     List<ProductWarehousingDTO> list(SearchDTO searchDTO);
